move atari offline examples to offline directory
