+++ examples/offline/atari_bcq.py
@@ -6,15 +6,17 @@ import pprint
 
 import numpy as np
 import torch
-from atari_network import QRDQN
-from atari_wrapper import wrap_deepmind
 from torch.utils.tensorboard import SummaryWriter
 
+from examples.atari.atari_network import DQN
+from examples.atari.atari_wrapper import wrap_deepmind
 from tianshou.data import Collector, VectorReplayBuffer
 from tianshou.env import ShmemVectorEnv
-from tianshou.policy import DiscreteCQLPolicy
+from tianshou.policy import DiscreteBCQPolicy
 from tianshou.trainer import offline_trainer
 from tianshou.utils import TensorboardLogger
+from tianshou.utils.net.common import ActorCritic
+from tianshou.utils.net.discrete import Actor
 
 
 def get_args():
@@ -22,12 +24,12 @@ def get_args():
     parser.add_argument("--task", type=str, default="PongNoFrameskip-v4")
     parser.add_argument("--seed", type=int, default=1626)
     parser.add_argument("--eps-test", type=float, default=0.001)
-    parser.add_argument("--lr", type=float, default=0.0001)
+    parser.add_argument("--lr", type=float, default=6.25e-5)
     parser.add_argument("--gamma", type=float, default=0.99)
-    parser.add_argument('--num-quantiles', type=int, default=200)
     parser.add_argument("--n-step", type=int, default=1)
-    parser.add_argument("--target-update-freq", type=int, default=500)
-    parser.add_argument("--min-q-weight", type=float, default=10.)
+    parser.add_argument("--target-update-freq", type=int, default=8000)
+    parser.add_argument("--unlikely-action-threshold", type=float, default=0.3)
+    parser.add_argument("--imitation-logits-penalty", type=float, default=0.01)
     parser.add_argument("--epoch", type=int, default=100)
     parser.add_argument("--update-per-epoch", type=int, default=10000)
     parser.add_argument("--batch-size", type=int, default=32)
@@ -67,7 +69,7 @@ def make_atari_env_watch(args):
     )
 
 
-def test_discrete_cql(args=get_args()):
+def test_discrete_bcq(args=get_args()):
     # envs
     env = make_atari_env(args)
     args.state_shape = env.observation_space.shape or env.observation_space.n
@@ -84,25 +86,38 @@ def test_discrete_cql(args=get_args()):
     torch.manual_seed(args.seed)
     test_envs.seed(args.seed)
     # model
-    net = QRDQN(*args.state_shape, args.action_shape, args.num_quantiles, args.device)
-    optim = torch.optim.Adam(net.parameters(), lr=args.lr)
-    # define policy
-    policy = DiscreteCQLPolicy(
-        net,
-        optim,
-        args.gamma,
-        args.num_quantiles,
-        args.n_step,
-        args.target_update_freq,
-        min_q_weight=args.min_q_weight
+    feature_net = DQN(
+        *args.state_shape, args.action_shape, device=args.device, features_only=True
     ).to(args.device)
+    policy_net = Actor(
+        feature_net,
+        args.action_shape,
+        device=args.device,
+        hidden_sizes=args.hidden_sizes,
+        softmax_output=False
+    ).to(args.device)
+    imitation_net = Actor(
+        feature_net,
+        args.action_shape,
+        device=args.device,
+        hidden_sizes=args.hidden_sizes,
+        softmax_output=False
+    ).to(args.device)
+    actor_critic = ActorCritic(policy_net, imitation_net)
+    optim = torch.optim.Adam(actor_critic.parameters(), lr=args.lr)
+    # define policy
+    policy = DiscreteBCQPolicy(
+        policy_net, imitation_net, optim, args.gamma, args.n_step,
+        args.target_update_freq, args.eps_test, args.unlikely_action_threshold,
+        args.imitation_logits_penalty
+    )
     # load a previous policy
     if args.resume_path:
         policy.load_state_dict(torch.load(args.resume_path, map_location=args.device))
         print("Loaded agent from: ", args.resume_path)
     # buffer
     assert os.path.exists(args.load_buffer_name), \
-        "Please run atari_qrdqn.py first to get expert's data buffer."
+        "Please run atari_dqn.py first to get expert's data buffer."
     if args.load_buffer_name.endswith('.pkl'):
         buffer = pickle.load(open(args.load_buffer_name, "rb"))
     elif args.load_buffer_name.endswith('.hdf5'):
@@ -116,7 +131,7 @@ def test_discrete_cql(args=get_args()):
 
     # log
     log_path = os.path.join(
-        args.logdir, args.task, 'cql',
+        args.logdir, args.task, 'bcq',
         f'seed_{args.seed}_{datetime.datetime.now().strftime("%m%d-%H%M%S")}'
     )
     writer = SummaryWriter(log_path)
@@ -164,4 +179,4 @@ def test_discrete_cql(args=get_args()):
 
 
 if __name__ == "__main__":
-    test_discrete_cql(get_args())
+    test_discrete_bcq(get_args())
