+++ tianshou/trainer/offline.py
@@ -27,7 +27,7 @@ def offline_trainer(
     reward_metric: Optional[Callable[[np.ndarray], np.ndarray]] = None,
     logger: BaseLogger = LazyLogger(),
     verbose: bool = True,
-    test_in_train: bool = True,
+    disable_test: bool = False,
 ) -> Dict[str, Union[float, str]]:
     """A wrapper for offline trainer procedure.
 
@@ -66,7 +66,7 @@ def offline_trainer(
     :param BaseLogger logger: A logger that logs statistics during updating/testing.
         Default to a logger that doesn't log anything.
     :param bool verbose: whether to print the information. Default to True.
-    :param bool test_in_train: whether to test in the training phase. Default to True.
+    :param bool disable_test: whether to run tests at all. Default to False.
 
     :return: See :func:`~tianshou.trainer.gather_info`.
     """
@@ -76,8 +76,9 @@ def offline_trainer(
     stat: Dict[str, MovAvg] = defaultdict(MovAvg)
     start_time = time.time()
     test_collector.reset_stat()
+    best_reward, best_reward_std = 0, 0
 
-    if test_in_train:
+    if not disable_test:
         test_result = test_episode(
             policy, test_collector, test_fn, start_epoch, episode_per_test, logger,
             gradient_step, reward_metric
@@ -101,7 +102,7 @@ def offline_trainer(
                 logger.log_update_data(losses, gradient_step)
                 t.set_postfix(**data)
         # test
-        if test_in_train:
+        if not disable_test:
             test_result = test_episode(
                 policy, test_collector, test_fn, epoch, episode_per_test, logger,
                 gradient_step, reward_metric
@@ -112,11 +113,15 @@ def offline_trainer(
                 if save_fn:
                     save_fn(policy)
         logger.save_data(epoch, 0, gradient_step, save_checkpoint_fn)
-        if verbose and test_in_train:
+        if verbose and not disable_test:
             print(
                 f"Epoch #{epoch}: test_reward: {rew:.6f} ± {rew_std:.6f}, best_rew"
                 f"ard: {best_reward:.6f} ± {best_reward_std:.6f} in #{best_epoch}"
             )
         if stop_fn and stop_fn(best_reward):
             break
+
+    if disable_test and save_fn:
+        save_fn(policy)
+
     return gather_info(start_time, None, test_collector, best_reward, best_reward_std)
